fix bug

+++ experiments/hrnet/compress_hrnet.py
@@ -186,7 +186,7 @@ def export_to_onnx(model, input_tensor, output_file_path):
     # input_names = [f"input_{i}" for i in range(len(input_tensors))]
     if torch.cuda.is_available():
         model.cuda().float()
-        input_tensor = input_tensor.cuda()
+        input_tensor = input_tensor.cuda().float()
     input_names = ["input"]
     with torch.no_grad():
         outputs = model(input_tensor)
